Move Curry.LanguageServer.Features.* to .Handlers.*
