+++ src/Curry/LanguageServer/Handlers.hs
@@ -4,5 +4,5 @@ import Curry.LanguageServer.Monad (LSM)
 import qualified Language.LSP.Server as S
 import qualified Language.LSP.Types as J
 
-handlers :: S.Flags -> S.Handlers LSM
-handlers flags = mconcat []
+handlers :: S.Handlers LSM
+handlers = mconcat []
